requirements.txt
yt-dlp>=2024.04.09
langchain==0.0.334
openai==1.7.1
pymongo>=4.6.3
python-dotenv==1.0.0
flask-cors==4.0.0
flask==2.3.2
PyPDF2==3.0.1
python-docx==1.1.0
openpyxl==3.1.2
pytz==2024.1
svix
